fix(windows): split capsule host and restore full viewport
